+++ src/association.spec.ts
@@ -1,7 +1,7 @@
 /* tslint:disable:completed-docs */
 import * as chai from 'chai';
 
-import { assoc, HAS_ONE } from './association';
+import { assoc, HAS_ONE, isLazyLoad } from './association';
 import { getAssociations } from './metadata';
 import { model, Model } from './model';
 
@@ -19,5 +19,18 @@ describe('@assoc', () => {
     let assocs = getAssociations(Entity);
 
     chai.assert.deepEqual(assocs['other'], { type: HAS_ONE, target: OtherEntity, readOnly: false });
+  });
+});
+
+describe('isLazyLoad', () => {
+  it('should return true for lambdas or functions returning a model', () => {
+    chai.should().equal(isLazyLoad(() => Entity), true);
+
+    // tslint:disable-next-line:only-arrow-functions
+    chai.should().equal(isLazyLoad(function() { return Entity; }), true);
+  });
+
+  it('should return false for a regular model', () => {
+    chai.should().equal(isLazyLoad(Entity), false);
   });
 });